added some sources
