+++ SoraStream/src/main/kotlin/com/hexated/SoraExtractor.kt
@@ -1,0 +1,103 @@
+package com.hexated
+
+import android.util.Log
+import com.lagradost.cloudstream3.APIHolder
+import com.lagradost.cloudstream3.SubtitleFile
+import com.lagradost.cloudstream3.apmap
+import com.lagradost.cloudstream3.app
+import com.lagradost.cloudstream3.utils.AppUtils
+import com.lagradost.cloudstream3.utils.ExtractorLink
+import com.lagradost.cloudstream3.utils.Qualities
+import com.lagradost.cloudstream3.utils.loadExtractor
+
+object SoraExtractor : SoraStream() {
+
+    suspend fun invokeLocalSources(
+        url: String,
+        subtitleCallback: (SubtitleFile) -> Unit,
+        callback: (ExtractorLink) -> Unit
+    ) {
+        val doc = app.get(
+            url,
+            headers = mapOf("User-Agent" to RandomUserAgent.getRandomUserAgent())
+        ).document
+        val script = doc.select("script").find { it.data().contains("\"sources\":[") }?.data()
+        val sourcesData = script?.substringAfter("\"sources\":[")?.substringBefore("],")
+        val subData = script?.substringAfter("\"subtitles\":[")?.substringBefore("],")
+
+        AppUtils.tryParseJson<List<Sources>>("[$sourcesData]")?.map { source ->
+            callback.invoke(
+                ExtractorLink(
+                    this.name,
+                    this.name,
+                    source.url ?: return@map null,
+                    "$mainServerAPI/",
+                    source.quality?.toIntOrNull() ?: Qualities.Unknown.value,
+                    isM3u8 = source.isM3U8,
+                    headers = mapOf("Origin" to mainServerAPI)
+                )
+            )
+        }
+
+        AppUtils.tryParseJson<List<Subtitles>>("[$subData]")?.map { sub ->
+            subtitleCallback.invoke(
+                SubtitleFile(
+                    sub.lang.toString(),
+                    sub.url ?: return@map null
+                )
+            )
+        }
+
+    }
+
+    suspend fun invokeTwoEmbed(
+        id: Int? = null,
+        season: Int? = null,
+        episode: Int? = null,
+        subtitleCallback: (SubtitleFile) -> Unit,
+        callback: (ExtractorLink) -> Unit
+    ) {
+        val url = if (season == null) {
+            "$twoEmbedAPI/embed/tmdb/movie?id=$id"
+        } else {
+            "$twoEmbedAPI/embed/tmdb/tv?id=$id&s=$season&e=$episode"
+        }
+        val document = app.get(url).document
+        val captchaKey =
+            document.select("script[src*=https://www.google.com/recaptcha/api.js?render=]")
+                .attr("src").substringAfter("render=")
+
+        document.select(".dropdown-menu a[data-id]").map { it.attr("data-id") }.apmap { serverID ->
+            val token = APIHolder.getCaptchaToken(url, captchaKey)
+            app.get(
+                "$twoEmbedAPI/ajax/embed/play?id=$serverID&_token=$token",
+                referer = url
+            ).parsedSafe<EmbedJson>()?.let { source ->
+                Log.i("hexated", "${source.link}")
+                loadExtractor(
+                    source.link ?: return@let null,
+                    twoEmbedAPI,
+                    subtitleCallback,
+                    callback
+                )
+            }
+        }
+    }
+
+    suspend fun invokeVidSrcSources(
+        id: Int? = null,
+        season: Int? = null,
+        episode: Int? = null,
+        subtitleCallback: (SubtitleFile) -> Unit,
+        callback: (ExtractorLink) -> Unit
+    ) {
+        val url = if (season == null) {
+            "$vidSrcAPI/embed/id=$id"
+        } else {
+            "$mainUrl/embed/$id/${season}-${episode}"
+        }
+
+        loadExtractor(url, null, subtitleCallback, callback)
+    }
+
+}
